[experiment] Replace vtable with trait object
There were 2 major differences between the two versions: 1) Memory layout and
2) Performance.

1) Memory layout

Since the `vtable` field of the `Bytes` struct is now a reference to a trait
object rather than an instance of a struct, the memory layout of it is affected
by the one of the reference to the trait object. For now, it adds 8 more bytes
to the `Bytes` struct: 32 -> 40. For whom wants to fix the size of the `Bytes`
struct, this change would not be a good choice since the memory layout of a
reference of a trait object is not stable and may change as described in
[the Rust reference](https://doc.rust-lang.org/reference/type-layout.html).

In the existing design, the `data` field of the `Bytes` struct could not be
moved into the trait object which the new `vtable` field represents to make the
type compact, since the `data` field could be replaced with another one at
runtime.

2) Performance

While the benchmark result shows almost the same results on most of the tests,
The `slice_empty`, `slice_short_from_arc`, and `split_off_and_drop` tests show
meaningful differences as below. They may come from the additional indirection
due to the trait object's vtable(needs analysis if needed, for now it is just a
guess).

- Before
```
test slice_empty          ... bench:       2,306 ns/iter (+/- 32)
test slice_short_from_arc ... bench:       9,773 ns/iter (+/- 63)
test split_off_and_drop   ... bench:      88,349 ns/iter (+/- 1,621)
```

- After
```
test slice_empty          ... bench:       2,645 ns/iter (+/- 21)
test slice_short_from_arc ... bench:      10,369 ns/iter (+/- 64)
test split_off_and_drop   ... bench:      88,896 ns/iter (+/- 1,013)
```